Implemented basic financial analysis concepts
Analysed and compared stock data of three car companies namely Tesla, Ford and GM. Along with this, it visualised data using various matplotlib modules with various parameters and tries to study relationships between them.

Calculated daily percentage change which basically informs you of your percent gain (or loss) if you bought the stock on day and then sold it the next day. These are very helpful in analyzing the volatility of the stock. If daily returns have a wide distribution, the stock is more volatile from one day to the next. Plotted them on a histogram to see which is most stable stock.
Calculated cumulative daily returns which is the aggregate amount an investment has gained or lost over time, independent of the period of time involved.  It takes daily returns into account. Cumulative return is computed relative to the day investment is made. If cumulative return is above one, you are making profits else you are in loss.